Refactor version history dialogs
